# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deskpilot",
    "summary": "Support agent for Acme Gear.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Exchange a password for a session.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Session",
        "description": "Trade a refresh token for a new pair.\n\nThis is the one endpoint authenticated by a cookie, so it is the one endpoint\na browser could be tricked into calling. SameSite=strict already stops that;\nthe double-submit check is a second lock on the same door, and costs one header.",
        "operationId": "refresh_session_api_auth_refresh_post",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "deskpilot_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deskpilot Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revoke the family and clear the cookie.\n\nNo CSRF check and no error when the token is unknown: being logged out against\nyour will is an annoyance, not a compromise, and a logout that reports whether\na token existed is an oracle (D-065).\n\nThe access token the client already holds stays valid until it expires, which\nis at most fifteen minutes. Killing it immediately would mean bumping\ntoken_version, and that signs the person out of every other device too -\nwhich is what \"sign out everywhere\" is for, not what \"sign out\" means. Short\naccess tokens are the answer, and they are already short.",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "deskpilot_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deskpilot Refresh"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Who the access token says you are.",
        "operationId": "me_api_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Open Ticket",
        "description": "Open a ticket and let the agent answer the first message.",
        "operationId": "open_ticket_api_tickets_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTicket"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "My Tickets",
        "description": "The tickets belonging to the signed-in customer.",
        "operationId": "my_tickets_api_tickets_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketSummary"
                  },
                  "title": "Response My Tickets Api Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/stream": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Open Ticket Streaming",
        "description": "Open a ticket and stream the agent's first answer as it is written.",
        "operationId": "open_ticket_streaming_api_tickets_stream_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTicket"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{reference}/replies/stream": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Reply Streaming",
        "description": "Add a message and stream the answer.",
        "operationId": "reply_streaming_api_tickets__reference__replies_stream_post",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reply"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{reference}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Read Ticket",
        "description": "One ticket and its conversation.",
        "operationId": "read_ticket_api_tickets__reference__get",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{reference}/replies": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Reply To Ticket",
        "description": "Add a message to a ticket and let the agent respond.",
        "operationId": "reply_to_ticket_api_tickets__reference__replies_post",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reply"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "description": "Liveness only. Says nothing about the database on purpose: a health\nendpoint that reports which dependency is down tells an attacker too.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalated": {
            "type": "boolean",
            "title": "Escalated"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "answer",
          "status",
          "category",
          "escalated"
        ],
        "title": "Answer",
        "description": "What one turn produced, for the caller that asked for it."
      },
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Identity": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "customer_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Email"
          }
        },
        "type": "object",
        "required": [
          "email",
          "role",
          "full_name",
          "customer_email"
        ],
        "title": "Identity"
      },
      "Message": {
        "properties": {
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "speaker",
          "text"
        ],
        "title": "Message"
      },
      "NewTicket": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject"
          },
          "message": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "message"
        ],
        "title": "NewTicket"
      },
      "Reply": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Reply"
      },
      "TicketCategory": {
        "type": "string",
        "enum": [
          "shipping",
          "return_or_refund",
          "warranty",
          "order_status",
          "product",
          "other"
        ],
        "title": "TicketCategory",
        "description": "What a ticket is about. Decided by the classifier, and only ever advisory."
      },
      "TicketDetail": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "subject",
          "status",
          "category",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "TicketDetail"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "open",
          "awaiting_customer",
          "escalated",
          "resolved"
        ],
        "title": "TicketStatus"
      },
      "TicketSummary": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "subject",
          "status",
          "category",
          "created_at",
          "updated_at"
        ],
        "title": "TicketSummary"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_at",
          "email",
          "role"
        ],
        "title": "TokenResponse",
        "description": "What the SPA keeps in memory. The refresh token is not in here."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}